LPD-26036 - Refactor the paths to follow the class naming pattern
