+++ modules/apps/object/object-test/src/testIntegration/resources/com/liferay/object/web/internal/object/definitions/portlet/action/test/dependencies/test-object-folder-1.json
